backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
passlib[bcrypt]
python-jose[cryptography]
pydantic
python-dotenv
